[ci] Specify repository and ref explicitly on checkout
For pull requests originating from forked repositories,
the main repository is used, and CI is triggered
on the intended target branch.
This update introduces the specification of the repository
and ref in the GitHub Actions workflow, ensuring that
CI is triggered on the correct branch.
There is no impact on builds from the main repository.

Signed-off-by: Mykhailo Lohvynenko <Mykhailo_Lohvynenko@epam.com>